move it from web-api-service to vertx-web-openapi
