--- vertx-web-api-service/src/test/java/io/vertx/ext/web/api/service/it/models/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(name="web-api-service", groupPackage = "io.vertx.ext.web.api.service.it.models")
-package io.vertx.ext.web.api.service.it.models;
-
-import io.vertx.codegen.annotations.ModuleGen;
