--- vertx-web-api-service/src/test/java/io/vertx/ext/web/api/service/it/services/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(name="web-api-service", groupPackage = "io.vertx.ext.web.api.service.it.services")
-package io.vertx.ext.web.api.service.it.services;
-
-import io.vertx.codegen.annotations.ModuleGen;
