+++ vertx-web-openapi/src/test/java/io/vertx/ext/web/openapi/it/models/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(name="web-api-service", groupPackage = "io.vertx.ext.web.openapi.it.models")
+package io.vertx.ext.web.openapi.it.models;
+
+import io.vertx.codegen.annotations.ModuleGen;
