+++ vertx-web-openapi/src/test/java/io/vertx/ext/web/openapi/it/services/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(name="web-api-service", groupPackage = "io.vertx.ext.web.openapi.it.services")
+package io.vertx.ext.web.openapi.it.services;
+
+import io.vertx.codegen.annotations.ModuleGen;
