+++ vertx-web-openapi/src/test/resources/openapi.json
